+++ pypi_upload.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+rm -rf build dist
+python setup.py sdist bdist_wheel
+twine upload -r pypi dist/python*